Use gaussian pyramid rather than simple resizes

+++ src/sampling.rs
@@ -1,5 +1,5 @@
 pub trait Downsample {
     type Output;
-    fn downsample(&self, scale: f64) -> Self::Output;
+    fn downsample(&self, scale: f32) -> Self::Output;
 }
